+++ chevah/keycert/tests/keydata.py
@@ -271,8 +271,6 @@ nqEsrLLZuOdPb2HuFoQsT5cd+rZsz19yQPHYUs5IgnPLzdWZAhUAl1TqdmlAG/b4
 nnVchGiO9sML8MM=
 -----END DSA PRIVATE KEY-----"""
 
-
-
 publicDSA_lsh = (
     "{KDEwOnB1YmxpYy1rZXkoMzpkc2EoMTpwNjU6AIbwTOSsZ7Bl7U1KyMNqV"
     "13Tu7yRAtTr70PVI3QnfrPumf2UzCgpL1ljbKxSfAi05XvrE/1vfCFAsFYXRZLhQy0pKDE6c"